# backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pointmarket_backend_internal_dtos.APIError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  pointmarket_backend_internal_dtos.AwardBadgeRequestDTO:
    properties:
      badge_id:
        type: integer
      user_id:
        type: integer
    required:
    - badge_id
    - user_id
    type: object
  pointmarket_backend_internal_dtos.BadgeDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      points_min:
        type: integer
      title:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.CreateBadgeRequestDTO:
    properties:
      description:
        type: string
      points_min:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
...
  pointmarket_backend_internal_dtos.ListUserBadgesResponseDTO:
    properties:
      total:
        type: integer
      user_badges:
        items:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UserBadgeDTO'
        type: array
    type: object
...
  pointmarket_backend_internal_dtos.NullData:
    type: object
...
  pointmarket_backend_internal_dtos.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginationMeta'
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.UpdateBadgeRequestDTO:
    properties:
      description:
        type: string
      points_min:
        type: integer
      title:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.UserBadgeDTO:
    properties:
      awarded_at:
        type: string
      badge_description:
        type: string
      badge_id:
        type: integer
      badge_title:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: PointMarket backend API documentation.
  title: PointMarket API
  version: "1.0"
paths:
...
  /badges:
    get:
      description: Default response is paginated list of badges. If `user_id` is provided,
        returns all badge awards for that user (admin-only).
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: User ID (admin-only; returns awards for that user)
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.ListUserBadgesResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: List badges
      tags:
      - badges
    post:
      consumes:
      - application/json
      parameters:
      - description: Badge
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateBadgeRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.BadgeDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Create badge
      tags:
      - badges
  /badges/{id}:
    delete:
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Delete badge
      tags:
      - badges
    get:
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.BadgeDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get badge by ID
      tags:
      - badges
    put:
      consumes:
      - application/json
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      - description: Badge
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpdateBadgeRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.BadgeDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Update badge
      tags:
      - badges
  /badges/{id}/award:
    post:
      consumes:
      - application/json
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      - description: Award
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.AwardBadgeRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Award badge
      tags:
      - badges
  /badges/{id}/revoke:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      - description: Revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.AwardBadgeRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Revoke badge
      tags:
      - badges
  /badges/awards:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pointmarket_backend_internal_dtos.UserBadgeDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: List badge awards
      tags:
      - badges
...
  /my-badges:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.ListUserBadgesResponseDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get my badges
      tags:
      - badges
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"